tools/tavisod/requirements.txt
google-cloud-secret-manager==2.12.6
